% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_match_summary.R
\name{vr_attack}
\alias{vr_attack}
\title{Generate attack table}
\usage{
vr_attack(x, team, by = "player", file_type = "indoor", style = "default")
}
\arguments{
\item{x}{datavolleyplays: the \code{plays} component of an object as returned by \code{datavolley::dv_read}}

\item{team}{string: team name}

\item{by}{string: "player" or "set"}

\item{file_type}{string: "indoor", "perana_indoor", "beach", "perana_beach"}

\item{style}{string: see \code{\link[=vr_match_summary]{vr_match_summary()}}}
}
\description{
Generate attack table
}
