% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vr_court_plots}
\alias{vr_court_plots}
\title{Court plots}
\usage{
vr_court_plots(x, ...)
}
\arguments{
\item{x}{datavolley or string: as returned by \code{datavolley::dv_read}, or the path to such a file}

\item{...}{: passed to downstream functions}
}
\value{
A ggplot object
}
\description{
Court plots
}
\examples{
vr_court_plots(dv_example_file())

}
