% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_match_summary.R
\name{vr_match_summary}
\alias{vr_match_summary}
\alias{vr_css}
\title{Generate single-page match summary report}
\usage{
vr_match_summary(
  x,
  outfile,
  vote = TRUE,
  format = "html",
  icon = NULL,
  css = vr_css(),
  remove_nonplaying = TRUE,
  shiny_progress = FALSE
)

vr_css()
}
\arguments{
\item{x}{datavolley or string: as returned by \code{datavolley::dv_read}, or the path to such a file}

\item{outfile}{string: path to file to produce (if not specified, will create a file in the temporary directory)}

\item{vote}{logical: include vote report component?}

\item{format}{string: "pdf", "png", or "html"}

\item{icon}{string: (optional) filename of icon image to use}

\item{css}{list: css specifications for some elements, giving (currently fairly limited) control over appearance. See the output of \code{\link{vr_css}} for an example. Note that some styling does not seem to be applied when exporting to PDF}

\item{remove_nonplaying}{logical: if \code{TRUE}, remove players from the team summaries that did not take to the court}

\item{shiny_progress}{logical: if \code{TRUE}, the report generation process will issue \code{shiny::setProgress()} calls. The call to \code{vr_match_summary} should therefore be wrapped in a \code{shiny::withProgress()} scope}
}
\value{
The path to the report file
}
\description{
Generate single-page match summary report
}
