% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vr_score_evplot}
\alias{vr_score_evplot}
\title{Score evolution plot}
\usage{
vr_score_evplot(
  x,
  home_colour = "darkblue",
  visiting_colour = "darkred",
  font_size = 12
)
}
\arguments{
\item{x}{datavolley or string: as returned by \code{datavolley::dv_read}, or the path to such a file}

\item{home_colour}{string: colour for home team}

\item{visiting_colour}{string: colour for visiting team}

\item{font_size}{scalar: font size}
}
\value{
A ggplot object
}
\description{
Score evolution plot
}
\examples{
vr_score_evplot(dv_example_file())

}
