% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_match_summary.R
\name{vr_serve}
\alias{vr_serve}
\title{Generate serve table}
\usage{
vr_serve(x, team, by = "player", refx, style = "default")
}
\arguments{
\item{x}{datavolleyplays: the \code{plays} component of an object as returned by \code{datavolley::dv_read}}

\item{team}{string: team name}

\item{by}{string: "player" or "set"}

\item{refx}{data.frame: see \code{\link[=vr_match_summary]{vr_match_summary()}}}

\item{style}{string: see \code{\link[=vr_match_summary]{vr_match_summary()}}}
}
\description{
Generate serve table
}
