% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_match_summary.R
\name{vr_vote}
\alias{vr_vote}
\title{Calculate vote}
\usage{
vr_vote(x, team)
}
\arguments{
\item{x}{datavolleyplays: the \code{plays} component of an object as returned by \code{datavolley::read_dv}}

\item{team}{string: team name}
}
\description{
Calculate vote
}
