% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_match_summary.R
\name{vr_points}
\alias{vr_points}
\title{Generate points table}
\usage{
vr_points(x, team, by = "player", vote = FALSE)
}
\arguments{
\item{x}{datavolleyplays: the \code{plays} component of an object as returned by \code{datavolley::read_dv}}

\item{team}{string: team name}

\item{by}{string: "player" or "set"}

\item{vote}{logical: if \code{TRUE}, include vote detail}
}
\description{
Generate points table
}
