% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vr_match_summary.R
\name{vr_reception}
\alias{vr_reception}
\title{Generate reception table}
\usage{
vr_reception(
  x,
  team,
  by = "player",
  refx,
  style = "default",
  file_type = "indoor"
)
}
\arguments{
\item{x}{datavolleyplays: the \code{plays} component of an object as returned by \code{datavolley::read_dv}}

\item{team}{string: team name}

\item{by}{string: "player" or "set"}

\item{refx}{data.frame: see \code{\link[=vr_match_summary]{vr_match_summary()}}}

\item{style}{string: see \code{\link[=vr_match_summary]{vr_match_summary()}}}

\item{file_type}{string: "indoor", "perana_indoor"}
}
\description{
Generate reception table
}
