% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vr_plot_markers}
\alias{vr_plot_markers}
\alias{vr_score_evplot}
\title{Score evolution plot}
\usage{
vr_plot_markers()

vr_score_evplot(
  x,
  with_summary = FALSE,
  markers = FALSE,
  home_colour = "darkblue",
  visiting_colour = "darkred",
  low_colour = "#800000",
  mid_colour = "#202020",
  high_colour = "#008000",
  font_size = 12
)
}
\arguments{
\item{x}{datavolley or string: as returned by \code{datavolley::dv_read}, or the path to such a file}

\item{with_summary}{logical: if \code{TRUE}, show team summary statistics. For beach, this is by end; for indoor summaries will be by set}

\item{markers}{logical or data.frame: either \code{FALSE} for no markers, \code{TRUE} to use the markers specified by \code{\link[=vr_plot_markers]{vr_plot_markers()}}, or a data.frame as returned by \code{\link[=vr_plot_markers]{vr_plot_markers()}}}

\item{home_colour}{string: colour for home team}

\item{visiting_colour}{string: colour for visiting team}

\item{low_colour, mid_colour, high_colour}{string: colours for low, mid, and high performance (used for \code{with_summary} only)}

\item{font_size}{scalar: font size}
}
\value{
A ggplot object
}
\description{
Score evolution plot
}
\examples{
vr_score_evplot(dv_example_file())

}
